requirements.txt
discord.py>=2.5.0
SQLAlchemy>=2.0.0
APScheduler>=3.10.4
python-dotenv>=1.0.1
httpx>=0.27.0
google-api-python-client>=2.167.0
google-auth>=2.39.0
google-auth-oauthlib>=1.2.2
google-auth-httplib2>=0.2.0
websockets>=15.0.1
